Update to mask region to be classified by the habitat and vegetation masks.

+++ 03_sen2_mangrove_cls/06_cls_sen2_scns/apply_scn_cls.py
@@ -4,6 +4,8 @@ import os
 import shutil
 import time
 import rsgislib
+import rsgislib.vectorutils
+import rsgislib.imagecalc
 import rsgislib.imageutils
 import rsgislib.classification.classxgboost
 
@@ -123,22 +125,69 @@ class ApplyXGBClass(PBPTQProcessTool):
             os.mkdir(self.params['tmp_dir'])
             time.sleep(1)
 
-        fileInfo = [rsgislib.imageutils.ImageBandInfo(self.params['sref_img'], 'sen2', [1, 2, 3, 4, 5, 6, 7, 8, 9, 10])]
-        outProbImg = os.path.join(self.params['tmp_dir'], "prob_cls_img.tif")
-        if os.path.exists(outProbImg):
-            os.remove(outProbImg)
-            time.sleep(1)
+        veg_msk_img = os.path.join(self.params['tmp_dir'], "{}_veg_msk.kea".format(self.params['scn_id']))
+        rsgislib.vectorutils.rasteriseVecLyr(self.params['gmw_veg_msk_vec'], self.params['gmw_veg_msk_lyr'],
+                                             self.params['vld_img'], veg_msk_img, gdalformat="KEA",
+                                             burnVal=1, datatype=rsgislib.TYPE_8UINT, vecAtt=None, vecExt=False,
+                                             thematic=True, nodata=0)
 
-        apply_xgboost_binary_classifier(self.params['cls_mdl_file'],
-                                        self.params['clrsky_img'], 1,
-                                        fileInfo, outProbImg, 'GTIFF',
-                                        outClassImg=self.params['out_cls_file'],
-                                        class_thres=5000, nthread=1)
+        hab_msk_img = os.path.join(self.params['tmp_dir'], "{}_hab_msk.kea".format(self.params['scn_id']))
+        rsgislib.vectorutils.rasteriseVecLyr(self.params['gmw_hab_msk_vec'], self.params['gmw_hab_msk_lyr'],
+                                             self.params['vld_img'], hab_msk_img, gdalformat="KEA",
+                                             burnVal=1, datatype=rsgislib.TYPE_8UINT, vecAtt=None, vecExt=False,
+                                             thematic=True, nodata=0)
+
+        hab_add_msk_img = os.path.join(self.params['tmp_dir'], "{}_hab_add_msk.kea".format(self.params['scn_id']))
+        rsgislib.vectorutils.rasteriseVecLyr(self.params['gmw_hab_msk_mng_add_vec'], self.params['gmw_hab_msk_mng_add_lyr'],
+                                             self.params['vld_img'], hab_add_msk_img, gdalformat="KEA",
+                                             burnVal=1, datatype=rsgislib.TYPE_8UINT, vecAtt=None, vecExt=False,
+                                             thematic=True, nodata=0)
+
+        hab_rm_msk_img = os.path.join(self.params['tmp_dir'], "{}_hab_rm_msk.kea".format(self.params['scn_id']))
+        rsgislib.vectorutils.rasteriseVecLyr(self.params['gmw_hab_msk_mng_rm_vec'], self.params['gmw_hab_msk_mng_rm_lyr'],
+                                             self.params['vld_img'], hab_rm_msk_img, gdalformat="KEA",
+                                             burnVal=1, datatype=rsgislib.TYPE_8UINT, vecAtt=None, vecExt=False,
+                                             thematic=True, nodata=0)
+
+        hab_msk_up_img = os.path.join(self.params['tmp_dir'], "{}_hab_msk_upd.kea".format(self.params['scn_id']))
+        bandDefns = []
+        bandDefns.append(rsgislib.imagecalc.BandDefn('hab', hab_msk_img, 1))
+        bandDefns.append(rsgislib.imagecalc.BandDefn('hab_add', hab_add_msk_img, 1))
+        bandDefns.append(rsgislib.imagecalc.BandDefn('hab_rm', hab_rm_msk_img, 1))
+        rsgislib.imagecalc.bandMath(hab_msk_up_img, '(hab_add==1)?1:(hab_rm==1)?0:hab', 'KEA', rsgislib.TYPE_8UINT, bandDefns)
+
+        cls_reg_msk = os.path.join(self.params['tmp_dir'], "{}_cls_msk.kea".format(self.params['scn_id']))
+        bandDefns = []
+        bandDefns.append(rsgislib.imagecalc.BandDefn('hab', hab_msk_up_img, 1))
+        bandDefns.append(rsgislib.imagecalc.BandDefn('clrsky', self.params['clrsky_img'], 1))
+        bandDefns.append(rsgislib.imagecalc.BandDefn('veg', veg_msk_img, 1))
+        rsgislib.imagecalc.bandMath(cls_reg_msk, '(hab==1) && (clrsky==1) && (veg==1)?1:0', 'KEA', rsgislib.TYPE_8UINT, bandDefns)
+
+        n_smpls = rsgislib.imagecalc.countPxlsOfVal(cls_reg_msk, vals=[1])
+        print(n_smpls)
+
+        if n_smpls[0] > 0:
+            fileInfo = [rsgislib.imageutils.ImageBandInfo(self.params['sref_img'], 'sen2', [1, 2, 3, 4, 5, 6, 7, 8, 9, 10])]
+            outProbImg = os.path.join(self.params['tmp_dir'], "prob_cls_img.tif")
+            if os.path.exists(outProbImg):
+                os.remove(outProbImg)
+                time.sleep(1)
+
+            apply_xgboost_binary_classifier(self.params['cls_mdl_file'],
+                                            cls_reg_msk, 1,
+                                            fileInfo, outProbImg, 'GTIFF',
+                                            outClassImg=self.params['out_cls_file'],
+                                            class_thres=5000, nthread=1)
+        else:
+            rsgislib.imageutils.createCopyImage(self.params['vld_img'], self.params['out_cls_file'], 1, 0, 'GTIFF', rsgislib.TYPE_8UINT)
+            
         if os.path.exists(self.params['tmp_dir']):
             shutil.rmtree(self.params['tmp_dir'])
 
     def required_fields(self, **kwargs):
-        return ["scn_id", "vld_img", "clrsky_img", "sref_img", "cls_mdl_file", "out_cls_file", "tmp_dir"]
+        return ["scn_id", "vld_img", "clrsky_img", "sref_img", "cls_mdl_file", "gmw_veg_msk_vec", "gmw_veg_msk_lyr",
+                "gmw_hab_msk_vec", "gmw_hab_msk_lyr", "gmw_hab_msk_mng_add_vec", "gmw_hab_msk_mng_add_lyr",
+                "gmw_hab_msk_mng_rm_vec", "gmw_hab_msk_mng_rm_lyr", "out_cls_file", "tmp_dir"]
 
     def outputs_present(self, **kwargs):
         files_dict = dict()
